+++ tipg/dbmodel.py
@@ -1,0 +1,295 @@
+"""tipg.dbmodel: database-domain models and the abstract Collection base."""
+
+import abc
+import datetime
+from typing import Any, Dict, List, Optional, TypedDict, Union
+
+from pydantic import BaseModel, model_validator
+
+from tipg.model import Extent
+
+# Links to geojson schema
+geojson_schema = {
+    "GEOMETRY": "https://geojson.org/schema/Geometry.json",
+    "POINT": "https://geojson.org/schema/Point.json",
+    "MULTIPOINT": "https://geojson.org/schema/MultiPoint.json",
+    "LINESTRING": "https://geojson.org/schema/LineString.json",
+    "MULTILINESTRING": "https://geojson.org/schema/MultiLineString.json",
+    "POLYGON": "https://geojson.org/schema/Polygon.json",
+    "MULTIPOLYGON": "https://geojson.org/schema/MultiPolygon.json",
+    "GEOMETRYCOLLECTION": "https://geojson.org/schema/GeometryCollection.json",
+}
+
+
+class Feature(TypedDict, total=False):
+    """Simple Feature model."""
+
+    type: str
+    # Geometry is either a dict or a str (wkt)
+    geometry: Optional[Union[Dict, str]]
+    properties: Optional[Dict]
+    id: Optional[Any]
+    bbox: Optional[List[float]]
+
+
+class ItemList(TypedDict):
+    """Items."""
+
+    items: List[Feature]
+    matched: Optional[int]
+    next: Optional[int]
+    prev: Optional[int]
+
+
+class Column(BaseModel):
+    """Model for database Column."""
+
+    name: str
+    type: str
+    description: Optional[str] = None
+    geometry_type: Optional[str] = None
+    srid: Optional[int] = None
+    bounds: Optional[List[float]] = None
+    mindt: Optional[str] = None
+    maxdt: Optional[str] = None
+
+    @model_validator(mode="before")
+    def sridbounds_default(cls, values):
+        """Set default bounds and srid when this is a function."""
+        if values.get("geometry_type"):
+            values["srid"] = values.get("srid", 4326)
+            values["bounds"] = values.get("bounds", [-180, -90, 180, 90])
+        return values
+
+    @property
+    def json_type(self) -> str:
+        """Return JSON field type."""
+        if self.type.endswith("[]"):
+            return "array"
+
+        if self.type in [
+            "smallint",
+            "integer",
+            "bigint",
+            "decimal",
+            "numeric",
+            "real",
+            "double precision",
+            "smallserial",
+            "serial",
+            "bigserial",
+            # Float8 is not a Postgres type name but is the name we give
+            # internally do Double Precision type
+            # ref: https://github.com/developmentseed/tipg/pull/60/files#r1011863866
+            "float8",
+        ]:
+            return "number"
+
+        if self.type.startswith("bool"):
+            return "boolean"
+
+        if any([self.type.startswith("json"), self.type.startswith("geo")]):
+            return "object"
+
+        return "string"
+
+    @property
+    def is_geometry(self) -> bool:
+        """Returns true if this property is a geometry column."""
+        return self.type in ("geometry", "geography")
+
+    @property
+    def is_datetime(self) -> bool:
+        """Returns true if this property is a datetime column."""
+        return self.type in ("timestamp", "timestamptz", "date")
+
+
+class Parameter(Column):
+    """Model for PostGIS function parameters."""
+
+    default: Optional[str] = None
+
+
+class Collection(BaseModel, metaclass=abc.ABCMeta):
+    """Collection Base Class."""
+
+    type: str
+    id: str
+    table: str
+    title: Optional[str] = None
+    description: Optional[str] = None
+    table_columns: List[Column] = []
+    properties: List[Column] = []
+    id_column: Optional[Column] = None
+    geometry_column: Optional[Column] = None
+    datetime_column: Optional[Column] = None
+    parameters: List[Parameter] = []
+
+    @property
+    def extent(self) -> Optional[Extent]:
+        """Return extent."""
+        extent: Dict[str, Any] = {}
+        if cols := self.geometry_columns:
+            if len(cols) == 1:
+                bbox = [cols[0].bounds]
+            else:
+                minx, miny, maxx, maxy = zip(*[col.bounds for col in cols])
+                bbox = [
+                    [min(minx), min(miny), max(maxx), max(maxy)],
+                    *[col.bounds for col in cols],
+                ]
+
+            extent["spatial"] = {
+                "bbox": bbox,
+                # The extent calculated in Pg is in WGS84 LON,LAT order
+                # so we use `CRS84` as CRS
+                "crs": "http://www.opengis.net/def/crs/OGC/1.3/CRS84",
+            }
+
+        if cols := [col for col in self.datetime_columns if col.mindt or col.maxdt]:
+            intervals = []
+            if len(cols) == 1:
+                if cols[0].mindt or cols[0].maxdt:
+                    intervals = [[cols[0].mindt, cols[0].maxdt]]
+
+            else:
+                mindt = [col.mindt for col in cols if col.mindt]
+                maxdt = [col.maxdt for col in cols if col.maxdt]
+                intervals = [
+                    [min(mindt), max(maxdt)],
+                    *[[col.mindt, col.maxdt] for col in cols if col.mindt or col.maxdt],
+                ]
+
+            if intervals:
+                extent["temporal"] = {"interval": intervals}
+
+        if extent:
+            return Extent(**extent)
+
+        return None
+
+    @property
+    def bounds(self) -> Optional[List[float]]:
+        """Return spatial bounds from collection extent."""
+        if self.extent and self.extent.spatial:
+            return self.extent.spatial.bbox[0]
+
+        return None
+
+    @property
+    def dt_bounds(self) -> Optional[List[str]]:
+        """Return temporal bounds from collection extent."""
+        if self.extent and self.extent.temporal:
+            return self.extent.temporal.interval[0]
+
+        return None
+
+    @property
+    def crs(self):
+        """Return crs of set geometry column."""
+        if self.geometry_column:
+            return f"http://www.opengis.net/def/crs/EPSG/0/{self.geometry_column.srid}"
+
+    @property
+    def geometry_columns(self) -> List[Column]:
+        """Return geometry columns."""
+        return [c for c in self.table_columns if c.is_geometry]
+
+    @property
+    def datetime_columns(self) -> List[Column]:
+        """Return datetime columns."""
+        return [c for c in self.table_columns if c.is_datetime]
+
+    def get_geometry_column(self, name: Optional[str] = None) -> Optional[Column]:
+        """Return the name of the first geometry column."""
+        if (not self.geometry_columns) or (name and name.lower() == "none"):
+            return None
+
+        if name is None:
+            return self.geometry_column
+
+        for col in self.geometry_columns:
+            if col.name == name:
+                return col
+
+        return None
+
+    def get_datetime_column(self, name: Optional[str] = None) -> Optional[Column]:
+        """Return the Column for either the passed in tstz column or the first tstz column."""
+        if not self.datetime_columns:
+            return None
+
+        if name is None:
+            return self.datetime_column
+
+        for col in self.datetime_columns:
+            if col.name == name:
+                return col
+
+        return None
+
+    def columns(self, properties: Optional[List[str]] = None) -> List[str]:
+        """Return table columns optionally filtered to only include columns from properties."""
+        if properties in [[], [""]]:
+            return []
+
+        cols = [
+            c.name for c in self.properties if c.type not in ["geometry", "geography"]
+        ]
+        if properties is None:
+            return cols
+
+        return [c for c in cols if c in properties]
+
+    def get_column(self, property_name: str) -> Optional[Column]:
+        """Return column info."""
+        for p in self.properties:
+            if p.name == property_name:
+                return p
+
+        return None
+
+    @property
+    def queryables(self) -> Dict:
+        """Return the queryables."""
+        if self.geometry_columns:
+            geoms = {
+                col.name: {"$ref": geojson_schema.get(col.geometry_type.upper(), "")}
+                for col in self.geometry_columns
+            }
+        else:
+            geoms = {}
+
+        props = {
+            col.name: {"name": col.name, "type": col.json_type}
+            for col in self.properties
+            if col.name not in geoms
+        }
+
+        return {**geoms, **props}
+
+    @abc.abstractmethod
+    async def features(self, *args, **kwargs) -> ItemList:
+        """Get Items."""
+        ...
+
+    @abc.abstractmethod
+    async def get_tile(self, *args, **kwargs) -> bytes:
+        """Get MVT Tile."""
+        ...
+
+
+class CollectionList(TypedDict):
+    """Collections."""
+
+    collections: List[Collection]
+    matched: Optional[int]
+    next: Optional[int]
+    prev: Optional[int]
+
+
+class Catalog(TypedDict):
+    """Internal Collection Catalog."""
+
+    collections: Dict[str, Collection]
+    last_updated: datetime.datetime